Cleanup suggestions from PR
